added PUT and exception

+++ src/main/java/com/cornershop/ecommerce/service/CategoryService.java
@@ -1,9 +1,12 @@
 package com.cornershop.ecommerce.service;
 
+import com.cornershop.ecommerce.exception.CategoryDeleteException;
 import com.cornershop.ecommerce.exception.CategoryDuplicateException;
 import com.cornershop.ecommerce.exception.CategoryNotFoundException;
 import com.cornershop.ecommerce.model.Category;
 import com.cornershop.ecommerce.repository.CategoryRepository;
+import com.cornershop.ecommerce.repository.ProductRepository;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +18,9 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
 
+    @Autowired
+    private ProductRepository productRepository;
+
     public Category createCategory(Category category) {
         Optional<Category> optionalCategory = categoryRepository.findCategoryByName(category.getName());
         if(optionalCategory.isPresent()) {
@@ -24,6 +30,10 @@ public class CategoryService {
     }
 
     public void deleteCategory(Long id) {
+        Long productCountOfCategory = productRepository.getProductCountOfCategoryId(id);
+        if(productCountOfCategory > 0) {
+            throw new CategoryDeleteException("you can not delete this category because category has " + productCountOfCategory + " products");
+        }
         categoryRepository.deleteById(id);
     }
 
@@ -32,6 +42,12 @@ public class CategoryService {
     }
 
     public List<Category> getAllCategoryList() {
-        return categoryRepository.findAll();
+        List<Category> categoryList = categoryRepository.findAll();
+        categoryList.sort(Comparator.comparingLong(Category::getId));
+        return categoryList;
+    }
+
+    public Category updateCategory(Category category) {
+        return categoryRepository.save(category);
     }
 }
